candidates and survey
